Inline upgrade engine into util_sa.c
