--- src/executables/system_metadata_upgrade.h
@@ -1,33 +1,0 @@
-/*
- *
- * Copyright 2016 CUBRID Corporation
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-
-/*
- * system_metadata_upgrade.h
- */
-
-#ifndef _SYSTEM_METADATA_UPGRADE_H_
-#define _SYSTEM_METADATA_UPGRADE_H_
-
-#include <stddef.h>
-
-extern int sysmeta_print_upgrade_scripts (int from_version, int to_version);
-extern int sysmeta_execute_upgrade_scripts (int from_version, int to_version);
-extern int sysmeta_execute_sql_buffer (char *buf, size_t len);
-
-#endif /* _SYSTEM_METADATA_UPGRADE_H_ */
